refactor: multiple improvements listed in the changelog
