+++ CHANGELOG.md
@@ -1,0 +1,22 @@
+# Change Log
+
+## [v0.2.1-alpha] - yyyy-mm-dd
+
+Change order ui list and info cluster selected
+
+### Added
+- Order list in each component
+- Added Makefile to run repetitive tasks
+### Changed
+- keys control over items 
+- Automatic size terminal
+### Fixed
+- **Documentation** error in keys for use
+- **error** linux/windows terminal to exit
+
+## [v0.2.0-alpha] - 2023-11-06
+First build
+
+### Added
+### Changed
+### Fixed
